Improve on code quality
