+++ src/main/java/duke/KBot/Launcher.java
@@ -8,6 +8,6 @@ import javafx.application.Application;
 public class Launcher {
     public static void main(String[] args) {
         TaskManager.loadLocalSavedTasks(); // checking if there are local files to load
-        Application.launch(Main.class, args);
+        Application.launch(KBot.class, args);
     }
 }
