+++ src/main/java/duke/memory/output.txt
@@ -1,5 +1,3 @@
-T | X | book return
-T | X | wack rahul
 T | X | sleep
 T | X | more sleep
 T | X | even more sleep
@@ -14,4 +12,4 @@ T | X | more selep
 T | X | more sleep
 T | X | more sleep
 T | X | more sleep
-T |   | more sleep
+T | X | more sleep
